+++ tests/test_stream.py
@@ -34,7 +34,7 @@ _ENCODER = json.JSONEncoder()  # type: ignore
 # pylint: disable=no-self-use
 
 
-class TestStream(object):
+class TestStream:
 
     @pytest.fixture(autouse=True)
     def chdir(self, tmpdir):
@@ -171,10 +171,10 @@ class TestStream(object):
                 }
             ]""")
 
-    class TestIterWrite(object):
+    class TestIterWrite:
         """Tests for the iterwrite object."""
 
-        class TestArray(object):
+        class TestArray:
             """Tests for array object."""
 
             @pytest.fixture(autouse=True)
@@ -202,7 +202,7 @@ class TestStream(object):
 
                 assert actual == list(range(5)) + ['a']
 
-        class TestObject(object):
+        class TestObject:
             """Tests for array object."""
 
             @pytest.fixture(autouse=True)
@@ -236,7 +236,7 @@ class TestStream(object):
                 assert actual == expected
 
 
-class TestObject(object):
+class TestObject:
 
     def test_init(self):
         with open('foo', 'w') as f:
@@ -253,13 +253,13 @@ class TestObject(object):
         with open('foo', 'r') as f:
             assert f.read() == '{}'
 
-    class TestWrite(object):
+    class TestWrite:
 
         @pytest.fixture(autouse=True)
         def chdir(self, tmpdir):
             tmpdir.chdir()
 
-        class TestWithoutIndent(object):
+        class TestWithoutIndent:
 
             @pytest.fixture(autouse=True)
             def chdir(self, tmpdir):
@@ -290,7 +290,7 @@ class TestObject(object):
                 with open('foo', 'r') as f:
                     assert f.read() == '{"foo": {"1": "bar"}}'
 
-        class TestWithIndent(object):
+        class TestWithIndent:
 
             @pytest.fixture(autouse=True)
             def chdir(self, tmpdir):
@@ -379,7 +379,7 @@ class TestObject(object):
                     }
                 }""")
 
-    class TestSubobject(object):
+    class TestSubobject:
         """Tests for the suboboject method."""
 
         @pytest.fixture(autouse=True)
@@ -436,7 +436,7 @@ class TestObject(object):
                         }
                     }""")
 
-        class TestNestedContextManager(object):
+        class TestNestedContextManager:
             """Test various nested configurations with context managers."""
 
             def test_inner(self):
@@ -481,7 +481,7 @@ class TestObject(object):
                     assert f.read() == \
                         '{"1": "bar", "ook": {"1": "bar"}, "foo": "bar"}'
 
-        class TestNested(object):
+        class TestNested:
             """Test various nested configurations."""
 
             def test_inner(self):
@@ -533,7 +533,7 @@ class TestObject(object):
                     assert f.read() == \
                         '{"1": "foo", "2": {"1": "bar"}, "3": "foo"}'
 
-    class TestSubarray(object):
+    class TestSubarray:
         """Tests for the subarray method."""
 
         @pytest.fixture(autouse=True)
@@ -590,7 +590,7 @@ class TestObject(object):
                         ]
                     }""")
 
-        class TestNestedContextManager(object):
+        class TestNestedContextManager:
             """Test various nested configurations with context managers."""
 
             def test_inner(self):
@@ -635,7 +635,7 @@ class TestObject(object):
                     assert f.read() == \
                         '{"1": "bar", "ook": [1], "foo": "bar"}'
 
-        class TestNested(object):
+        class TestNested:
             """Test various nested configurations."""
 
             def test_inner(self):
@@ -686,7 +686,7 @@ class TestObject(object):
                 with open('foo', 'r') as f:
                     assert f.read() == '{"1": "foo", "2": [1], "3": "foo"}'
 
-    class TestClose(object):
+    class TestClose:
 
         @pytest.fixture(autouse=True)
         def chdir(self, tmpdir):
@@ -732,7 +732,7 @@ class TestObject(object):
                 with pytest.raises(jsonstreams.StreamClosedError):
                     test.subobject('foo')
 
-    class TestWriteToParent(object):
+    class TestWriteToParent:
         """Tests for writing to the parent with a subobject open."""
 
         @pytest.fixture(autouse=True)
@@ -753,7 +753,7 @@ class TestObject(object):
                         with pytest.raises(jsonstreams.ModifyWrongStreamError):
                             a.write('foo', 'bar')
 
-    class TestIterWrite(object):
+    class TestIterWrite:
         """Tests for the iterwrite object."""
 
         @pytest.fixture(autouse=True)
@@ -840,7 +840,7 @@ class TestObject(object):
             assert actual == expected
 
 
-class TestArray(object):
+class TestArray:
 
     @pytest.fixture(autouse=True)
     def chdir(self, tmpdir):
@@ -861,13 +861,13 @@ class TestArray(object):
         with open('foo', 'r') as f:
             assert f.read() == '[]'
 
-    class TestWrite(object):
+    class TestWrite:
 
         @pytest.fixture(autouse=True)
         def chdir(self, tmpdir):
             tmpdir.chdir()
 
-        class TestWithoutIndent(object):
+        class TestWithoutIndent:
 
             @pytest.fixture(autouse=True)
             def chdir(self, tmpdir):
@@ -898,7 +898,7 @@ class TestArray(object):
                 with open('foo', 'r') as f:
                     assert f.read() == '[{"1": "bar"}]'
 
-        class TestWithIndent(object):
+        class TestWithIndent:
 
             @pytest.fixture(autouse=True)
             def chdir(self, tmpdir):
@@ -979,7 +979,7 @@ class TestArray(object):
                     }
                 ]""")
 
-    class TestSubobject(object):
+    class TestSubobject:
 
         @pytest.fixture(autouse=True)
         def chdir(self, tmpdir):
@@ -1035,7 +1035,7 @@ class TestArray(object):
                         }
                     ]""")
 
-        class TestNestedContextManager(object):
+        class TestNestedContextManager:
             """Test various nested configurations with context managers."""
 
             def test_inner(self):
@@ -1079,7 +1079,7 @@ class TestArray(object):
                 with open('foo', 'r') as f:
                     assert f.read() == '[1, {"1": "bar"}, 2]'
 
-        class TestNested(object):
+        class TestNested:
             """Test various nested configurations."""
 
             def test_inner(self):
@@ -1131,7 +1131,7 @@ class TestArray(object):
                 with open('foo', 'r') as f:
                     assert f.read() == '[1, {"1": "bar"}, 2]'
 
-    class TestSubarray(object):
+    class TestSubarray:
 
         @pytest.fixture(autouse=True)
         def chdir(self, tmpdir):
@@ -1187,7 +1187,7 @@ class TestArray(object):
                         ]
                     ]""")
 
-        class TestNestedContextManager(object):
+        class TestNestedContextManager:
             """Test various nested configurations with context managers."""
 
             def test_inner(self):
@@ -1231,7 +1231,7 @@ class TestArray(object):
                 with open('foo', 'r') as f:
                     assert f.read() == '[1, ["bar"], 2]'
 
-        class TestNested(object):
+        class TestNested:
             """Test various nested configurations."""
 
             def test_inner(self):
@@ -1283,7 +1283,7 @@ class TestArray(object):
                 with open('foo', 'r') as f:
                     assert f.read() == '[1, [1], 2]'
 
-    class TestClose(object):
+    class TestClose:
 
         @pytest.fixture(autouse=True)
         def chdir(self, tmpdir):
@@ -1329,7 +1329,7 @@ class TestArray(object):
                 with pytest.raises(jsonstreams.StreamClosedError):
                     test.subobject()
 
-    class TestWriteToParent(object):
+    class TestWriteToParent:
         """Tests for writing to the parent with a subobject open."""
 
         @pytest.fixture(autouse=True)
@@ -1350,7 +1350,7 @@ class TestArray(object):
                         with pytest.raises(jsonstreams.ModifyWrongStreamError):
                             a.write('foo')
 
-    class TestIterWrite(object):
+    class TestIterWrite:
         """Tests for the iterwrite object."""
 
         @pytest.fixture(autouse=True)