+++ jsonstreams/__init__.py
@@ -88,7 +88,7 @@ __all__ = (
     'Type',
 )
 
-__version__ = '0.4.3'
+__version__ = '0.5.0'
 
 
 class JSONStreamsError(Exception):