+++ docs/source/changes.rst
@@ -4,6 +4,11 @@ Changes
 next
 ------
 
+New Features
+
+- Drop support for python 2
+- Drop six requirement
+
 Bug Fixes:
 
 - Fix a bug that prevents the Stream class from producing proper compact JSON output
@@ -12,6 +17,7 @@ Bug Fixes:
   (`#23 <https://github.com/dcbaker/jsonstreams/issues/23>`_)
 - flush the fd before closing the Stream
   (`#24 <https://github.com/dcbaker/jsonstreams/issues/24>`_)
+
 
 0.4.3
 ------