+++ docs/source/changes.rst
@@ -1,7 +1,7 @@
 Changes
 =======
 
-next
+0.5.0
 ------
 
 Bug Fixes: